+++ axelor-sale/src/main/java/com/axelor/apps/sale/service/saleorder/SaleOrderService.java
@@ -48,20 +48,21 @@ public interface SaleOrderService {
 
   /**
    * Check modified confirmed order before saving it.
-   * @param saleOrder
-   * @param saleOrderView
-   * @throws AxelorException
-   */
-  void checkModifiedConfirmedOrder(SaleOrder saleOrder, SaleOrder saleOrderView) throws AxelorException;
-
-  /**
-   * Validate changes.
    *
    * @param saleOrder
    * @param saleOrderView
    * @throws AxelorException
    */
-  void validateChanges(SaleOrder saleOrder, SaleOrder saleOrderView) throws AxelorException;
+  void checkModifiedConfirmedOrder(SaleOrder saleOrder, SaleOrder saleOrderView)
+      throws AxelorException;
+
+  /**
+   * Validate changes.
+   *
+   * @param saleOrder
+   * @throws AxelorException
+   */
+  void validateChanges(SaleOrder saleOrder) throws AxelorException;
 
   /**
    * Sort detail lines by sequence.