+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - MANUF. ORDER : empty stockmoves with realizeBtn and finishPartBtn
 - CUSTOMER DELIVERY : editable deliveryCondition
 - CRM : Addition of 'freeText' boolean in LostReason and 'lostReasonStr' in lead and opportunity.
+- Sale Order: if config is activated, create stock move after modifications validation.
 
 ## Bug Fixes
 - StockMoveLine: Fix wrong price values in declaration of exchanges.